Implement InMemoryUserDetailsManager in security configuration
The update involves modifying the ProjectSecurityConfig.java file in order to set up InMemoryUserDetailsManager for user authentication. This involved adding predefined user details for an 'admin' and a 'user', along with their respective authorities. It also includes the removal of the previously hardcoded user credentials in the application.yml file, making the system more secure and robust.